move constants out of seed.mts
