+++ supabase/seed/clientsSeed.mts
@@ -1,0 +1,31 @@
+export const possibleDietaryRequirements = [
+    "Gluten Free",
+    "Dairy Free",
+    "Vegetarian",
+    "Vegan",
+    "Pescatarian",
+    "Halal",
+    "Diabetic",
+    "Nut Allergy",
+    "Seafood Allergy",
+    "No Bread",
+    "No Pasta",
+    "No Rice",
+    "No Pork",
+    "No Beef",
+];
+
+export const possibleFeminineProducts = ["Tampons", "Pads", "Incontinence Pads"];
+
+export const possiblePets = ["Cat", "Dog"];
+
+export const possibleOtherItems = [
+    "Garlic",
+    "Ginger",
+    "Chillies",
+    "Spices",
+    "Hot Water Bottles",
+    "Blankets",
+];
+
+export const possibleParcelPostCodes = ["E1 6AA", "E1 6AD", "E1 6AG", "CB2 3JU", "CB24 4RT", "CB8 9LJ"];
