+++ supabase/seed/collectionCentresSeed.mts
@@ -1,0 +1,12 @@
+export const collectionCentres = [{
+    name: "Brixton Hill - Methodist Church",
+    acronym: "BH-MC",
+},
+{
+    name: "Clapham - St Stephens Church",
+    acronym: "CLP-SC",
+},
+{
+    name: "Delivery",
+    acronym: "DLVR",
+}]
