+++ supabase/seed/eventsSeed.mts
@@ -1,0 +1,20 @@
+export const eventNamesWithNumberData = ["Shipping Labels Downloaded"];
+
+export const eventNamesWithNoData = [
+    "No Status",
+    "Request Denied",
+    "Pending More Info",
+    "Called and Confirmed",
+    "Called and No Response",
+    "Ready to Dispatch",
+    "Received by Centre",
+    "Collection Failed",
+    "Parcel Collected",
+    "Delivered",
+    "Delivery Failed",
+    "Delivery Cancelled",
+    "Fulfilled with Trussell Trust",
+    "Shopping List Downloaded",
+    "Out for Delivery",
+    "Request Deleted",
+];
