+++ supabase/seed/listsHotelsSeed.mts
@@ -1,0 +1,13 @@
+export const listsHotels = [{
+    itemName: "Canned Hot Dog",
+    quantityFor1: "1",
+    quantityFor2: "2",
+    quantityFor3: "3",
+    quantityFor4: "4",
+    quantityFor5: "5",
+    quantityFor6: "6",
+    quantityFor7: "7",
+    quantityFor8: "8",
+    quantityFor9: "9",
+    quantityFor10: "10",
+}]
