+++ supabase/seed/packingSlotsSeed.mts
@@ -1,0 +1,20 @@
+export const packingSlots = [{
+    name: "AM",
+    isShown: true,
+    order: 1,
+},
+{
+    name: "PM",
+    isShown: true,
+    order: 2,
+},
+{
+    name: "Slot 1",
+    isShown: false,
+    order: 3,
+},
+{
+    name: "Slot 2",
+    isShown: false,
+    order: 4,
+}]
