+++ supabase/seed/websiteDataSeed.mts
@@ -1,0 +1,8 @@
+export const websiteData = [{
+    name: "lists_text",
+    value:
+        "Space is valuable! Please don't leave boxes half empty - pack efficiently!\n" +
+        "BOXES MUST BE PACKED FLAT SO THAT THEY CAN BE STACKED. Do not leave items sticking out of the top.\n" +
+        "We do have a selection of 'free from' goods as well as vegan and halal products. " +
+        "If you're uncertain about any additional dietary needs, please speak to one of the team.",
+}]
